refactored snapshot related tests to a new file
